refactor: clean up api, move files
